requirements.txt
Flask
psycopg2-binary
opencv-python-headless
numpy
